Implement `--show` option
Closes #6.

Display the LED bargraph in the terminal. Supports colors, blinking,
etc. Can either show the value/range that was just set (`--show`) or
used alone as a command (`show`) to read the current value/range from
the LED bargraph and display it.

* Uses new `0.3.0` version of `HT16K33` library.
* Renamed option `--as-is` to be `--no-init`, now correctly skips
initialization of the bargraph if desired (avoids flicker when updating
an already-active display).
* Added `--verbose` logging option for `INFO` log level.
* Added docs and testing.

Also, disable CI for `*-musl` platforms, as they fail with a
missing `proc_macro` dependency failure due to not being able to load
dynamic libraries. It looks like this is introduced from `serde_derive`?